requirements.txt
amqp==2.6.0
anyjson==0.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
billiard==3.6.3.0
celery==4.4.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
dwebsocket-0.5.12
Django==2.2.0
django-celery-beat==2.0.0
django-cors-headers==3.3.0
django-picklefield==3.0.1
django-redis-sessions==0.6.1
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
eventlet==0.25.2
flower==0.9.4
funky==0.0.2
future==0.18.2
gitdb2==3.0.2
GitPython==3.0.8
greenlet==0.4.16
humanize==0.5.1
idna==2.8
importlib-metadata==1.6.1
itypes==1.1.0
Jinja2==2.11.1
jsonfield==3.1.0
kombu==4.6.10
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
pysnowflake==0.1.3
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.10.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
six==1.15.0
smmap2==2.0.5
sqlparse==0.3.0
toml==0.10.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.2.5
xlrd==1.2.0
zipp==3.1.0
